+++ packages/expo-modules-autolinking/CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - [Android] Introduced the Gradle plugin to improve the autolinking setup. ([#33402](https://github.com/expo/expo/pull/33402) by [@lukmccall](https://github.com/lukmccall))
 - Drop `fs-extra` in favor of `fs`. ([#35036](https://github.com/expo/expo/pull/35036) by [@kitten](https://github.com/kitten))
+- [Android] Added `publication` configuration.
 
 ## 2.0.4 - 2024-12-10